moved components into correct folders
