--- src/TeamColor.java
@@ -1,3 +1,0 @@
-public enum TeamColor {
-    RED, YELLOW, BLUE, GREEN
-}
