+++ src/test/java/com/game/components/DieTest.java
@@ -1,3 +1,5 @@
+package test.java.com.game.components;
+
 import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
@@ -9,6 +11,6 @@ public class DieTest {
     public void testRollDice(){
         int result = die.rollDice();
         System.out.println(die);
-        assertTrue("Roll result should be between 1 and 6", result >= 1 && result <= 6);
+        Assert.assertTrue("Roll result should be between 1 and 6", result >= 1 && result <= 6);
     }
 }
